Factored out repl part into firebase functions:terminal
