+++ lib/callableFunction.js
@@ -1,5 +1,5 @@
 module.exports = class callableFunction {
-	constructor(name, controller, https) {
+	constructor(name, controller) {
 		this.name = name;
 		this.controller = controller;
